requirements.txt
django
gunicorn
django-heroku
djangorestframework==3.9.0
